代码调整

+++ cloud-management/src/main/java/cn/roamblue/cloud/management/servcie/GuestService.java
@@ -199,7 +199,7 @@ public class GuestService {
     }
     @Lock(RedisKeyUtil.GLOBAL_LOCK_KEY)
     @Transactional(rollbackFor = Exception.class)
-    public ResultUtil<GuestModel> modifyGuest(int guestId, String description, String busType, int cpu, int memory) {
+    public ResultUtil<GuestModel> modifyGuest(int guestId, String description, String busType, int cpu, long memory) {
         GuestEntity guest = this.guestMapper.selectById(guestId);
         if (guest.getStatus() != Constant.GuestStatus.STOP) {
             throw new CodeException(ErrorCode.VM_NOT_STOP, "请首先停止系统");
